Multiplayer works! Also tweaked UI
Fixed multiplayer not syncing map and day/night cycle. I did this by making a world prefab which is a GameObject containing the sky (sun, moon, stars) and terrain. World is a NetworkObject and is instantiated when a player begins hosting. Moon and sun are synced by attaching network transform components. Terrain is synced by having the terrain generation seed as a network variable which is set by the host and used by all clients.